sidecar: move sidecar logic to new file, properly handle retransmission
In this commit, we move most of the sidecar logic to a new file, and also tweak each of the main state machines to be able to properly handle retransmission events by both sides. When the provider restarts, it'll send over a plain offered ticket, which'll prompt the receiver to re-send their registered ticket. When the receiver restarts, it'll re-send its registered ticket, which'll prompt the provider to re-send their ordered ticket.