auctioneer+rpc: force a reconnection attempt on unknown error
As is, we only attempt to reconnect if we receive a nice io.EOF error
from the gRPC client, which is usually triggered when the server
shutsdown. However in the wild, possibly due to a load balancer and/or
reverse proxy bug, we encounter another error which as is, doesn't
trigger the usual reconnection logic.

In this commit, we now log any unknown error, but then also attempt to
reconnect again by sending the normal "server shutdown" message.

As is, if this doesn't work after the max number of retries, we won't
retry any longer, but perhaps we should still attempt to retry in the
background.